move theme to utils folder
